Create MenuItems components, create isSmallScreen utils, rename components, remove SmallMainMenuItem component
