+++ src/components/features/MenuDrawer.tsx
@@ -1,14 +1,12 @@
 import MainMenu from '@/components/features/MainMenu'
+import { isSmallScreenUtils } from '@/utils/isSmallScreenUtils'
 import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew'
 import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos'
 import { Box } from '@mui/system'
 import { useState } from 'react'
 
-interface MenuDrawerProps {
-  isSmallScreen: boolean
-}
-
-export default function MenuDrawer({ isSmallScreen }: MenuDrawerProps) {
+export default function MenuDrawer() {
+  const isSmallScreen: boolean = isSmallScreenUtils()
   const [isHideMenu, setIsHideMenu] = useState(false)
 
   function hideMenuHandler() {
