+++ src/components/features/Toolbar.tsx
@@ -2,12 +2,10 @@ import { Box } from '@mui/material'
 import ProfileMenu from '@/components/features/ProfileMenu'
 import SmallMainMenu from '@/components/features/SmallMainMenu'
 import LanguageSwitcher from '@/components/features/LanguageSwitcher'
+import { isSmallScreenUtils } from '@/utils/isSmallScreenUtils'
 
-interface ToolbarProps {
-  isSmallScreen: boolean
-}
-
-export default function Toolbar({ isSmallScreen }: ToolbarProps) {
+export default function Toolbar() {
+  const isSmallScreen: boolean = isSmallScreenUtils()
   return (
     <Box
       component="nav"
