+++ src/components/shared/ShowHideFunctionality.tsx
@@ -1,6 +1,5 @@
 import { IconButton, TextField } from '@mui/material'
 import InputAdornment from '@mui/material/InputAdornment'
-
 import React from 'react'
 import { useTranslation } from 'react-i18next'
