+++ src/utils/isSmallScreenUtils.ts
@@ -1,0 +1,6 @@
+import { useMediaQuery } from '@mui/material'
+
+export function isSmallScreenUtils(): boolean {
+  const isSmallScreen = useMediaQuery('(max-width: 800px)')
+  return isSmallScreen
+}
